adding new module and tests.
