+++ src/_gettsim/functions.py
@@ -29,5 +29,6 @@ from _gettsim.transfers.kinderzuschl.kinderzuschl import *  # noqa: F403
 from _gettsim.transfers.kinderzuschl.kinderzuschl_eink import *  # noqa: F403
 from _gettsim.transfers.kinderzuschl.kost_unterk import *  # noqa: F403
 from _gettsim.transfers.rente import *  # noqa: F403
+from _gettsim.transfers.unterhalt import *  # noqa: F403
 from _gettsim.transfers.unterhaltsvors import *  # noqa: F403
 from _gettsim.transfers.wohngeld import *  # noqa: F403
