+++ src/_gettsim/transfers/unterhalt.py
@@ -1,0 +1,31 @@
+"""This module provides functions to compute alimony payments (Unterhalt)."""
+
+
+def kind_unterh_netto_m(
+    kind_unterh_brutto_m: float,
+    kindergeld_m: float,
+    unterhalt_params: dict,
+    kind: bool,
+) -> float:
+    """Monthly actual child alimony payments to be received after deductions.
+
+    Parameters
+    ----------
+    kind_unterh_brutto_m
+        See basic input variable :ref:`kind_unterh_brutto_m <kind_unterh_brutto_m>`.
+    kindergeld_m
+        See :func:`kindergeld_m`.
+    unterhalt_params
+        See params documentation :ref:`unterhalt_params <unterhalt_params>`.
+    kind
+        See basic input variable :ref:`kind <kind>`.
+
+    Returns
+    -------
+    """
+    if kind:
+        abzugsrate = unterhalt_params["abzugsrate_kindergeld"]["kind"]
+    else:
+        abzugsrate = unterhalt_params["abzugsrate_kindergeld"]["erwachsener"]
+
+    return kind_unterh_brutto_m - abzugsrate * kindergeld_m
